--- bench.c
@@ -1,845 +1,0 @@
-/*
-	Copyright (C) 2014-2021 Igor van den Hoven ivdhoven@gmail.com
-*/
-
-/*
-	Permission is hereby granted, free of charge, to any person obtaining
-	a copy of this software and associated documentation files (the
-	"Software"), to deal in the Software without restriction, including
-	without limitation the rights to use, copy, modify, merge, publish,
-	distribute, sublicense, and/or sell copies of the Software, and to
-	permit persons to whom the Software is furnished to do so, subject to
-	the following conditions:
-
-	The above copyright notice and this permission notice shall be
-	included in all copies or substantial portions of the Software.
-
-	The person recognize Mars as a free planet and that no Earth-based
-	government has authority or sovereignty over Martian activities.
-
-	THE SOFTWARE IS PROVIDED "AS IS", WITHOUT WARRANTY OF ANY KIND,
-	EXPRESS OR IMPLIED, INCLUDING BUT NOT LIMITED TO THE WARRANTIES OF
-	MERCHANTABILITY, FITNESS FOR A PARTICULAR PURPOSE AND NONINFRINGEMENT.
-	IN NO EVENT SHALL THE AUTHORS OR COPYRIGHT HOLDERS BE LIABLE FOR ANY
-	CLAIM, DAMAGES OR OTHER LIABILITY, WHETHER IN AN ACTION OF CONTRACT,
-	TORT OR OTHERWISE, ARISING FROM, OUT OF OR IN CONNECTION WITH THE
-	SOFTWARE OR THE USE OR OTHER DEALINGS IN THE SOFTWARE.
-*/
-
-/*
-	quadsort 1.1.3.3
-
-	To compile use:
-
-	gcc -O3 bench.c
-
-	or
-
-	g++ -O3 -w -fpermissive bench.c
-*/
-
-#include <stdlib.h>
-#include <stdio.h>
-#include <string.h>
-#include <sys/time.h>
-#include <time.h>
-#include <assert.h>
-#include <errno.h>
-#include "quadsort.h"
-
-//#define cmp(a,b) (*(a) > *(b))
-
-typedef int CMPFUNC (const void *a, const void *b);
-
-typedef void SRTFUNC(void *array, size_t nmemb, size_t size, CMPFUNC *ignore);
-
-// benchmarking utilities
-
-
-// Must prevent inlining so the benchmark is fair.
-// Remove __attribute__ ((noinline)) and counter++ for full throttle.
-
-unsigned int counter;
-
-__attribute__ ((noinline)) int cmp_int(const void * a, const void * b)
-{
-	counter++;
-
-	return *(int *) a - *(int *) b;
-}
-
-__attribute__ ((noinline)) int cmp_stable(const void * a, const void * b)
-{
-	counter++;
-
-	return *(int *) a / 1000 - *(int *) b / 1000;
-}
-
-__attribute__ ((noinline)) int cmp_long(const void * a, const void * b)
-{
-	counter++;
-
-	if (*(long long *) a < *(long long *) b)
-	{
-		return -1;
-	}
-
-	return *(long long *) a > *(long long *) b;
-}
-
-int cmp_str(const void * a, const void * b)
-{
-	return strcmp(*(const char **) a, *(const char **) b);
-}
-
-int cmp_float(const void * a, const void * b)
-{
-	return *(float *) a - *(float *) b;
-}
-
-
-long long utime()
-{
-	struct timeval now_time;
-
-	gettimeofday(&now_time, NULL);
-
-	return now_time.tv_sec * 1000000LL + now_time.tv_usec;
-}
-
-
-void test_sort(void *array, void *unsorted, void *valid, int minimum, int maximum, int samples, int repetitions, SRTFUNC *srt, const char *name, const char *desc, size_t size, CMPFUNC *cmpf)
-{
-	double comps;
-	long long start, end, total, best, average;
-	size_t rep, sam, max;
-	long long *ptla = array, *ptlv = valid;
-	int *pta = array, *ptv = valid, cnt;
-
-	best = average = comps = start = 0;
-
-	if (maximum == 10 && size == sizeof(int))
-	{
-		memcpy(array, unsorted, maximum * size);
-
-		printf("\e[1;32m in: \e[1;31m%10d %10d %10d %10d %10d %10d %10d %10d %10d %10d\e[0m\n", pta[0], pta[1], pta[2], pta[3], pta[4], pta[5], pta[6], pta[7], pta[8], pta[9]);
-	}
-
-	for (sam = 0 ; sam < samples ; sam++)
-	{
-		total = 0;
-
-		max = minimum;
-
-		if (repetitions > 100)
-		{
-			start = utime();
-		}
-
-		for (rep = 0 ; rep < repetitions ; rep++)
-		{
-			memcpy(array, unsorted, max * size);
-
-			counter = 0;
-
-			if (repetitions <= 100)
-			{
-				start = utime();
-			}
-
-			switch (*name)
-			{
-				case 'q':
-					if (name[1] == 'u')
-					{
-						quadsort(array, max, size, cmpf);
-					}
-					else
-					{
-						qsort(array, max, size, cmpf);
-					}
-					break;
-
-				default:
-					return;
-			}
-
-			if (repetitions <= 100)
-			{
-				end = utime();
-
-				total += end - start;
-			}
-			comps += counter;
-
-			if (minimum < maximum)
-			{
-				max++;
-
-				if (max > maximum)
-				{
-					max = minimum;
-				}
-			}
-		}
-
-		if (repetitions > 100)
-		{
-			end = utime();
-
-			total = end - start;
-		}
-
-		if (!best || total < best)
-		{
-			best = total;
-		}
-		average += total;
-	}
-
-	if (maximum == 10 && size == sizeof(int))
-	{
-		printf("\e[1;32mout: \e[1;31m%10d %10d %10d %10d %10d %10d %10d %10d %10d %10d\e[0m\n", pta[0], pta[1], pta[2], pta[3], pta[4], pta[5], pta[6], pta[7], pta[8], pta[9]);
-	}
-
-	if (repetitions == 0)
-	{
-		return;
-	}
-
-	comps /= samples * repetitions;
-	average /= samples;
-
-	if (cmpf == cmp_stable)
-	{
-		for (cnt = 1 ; cnt < maximum ; cnt++)
-		{
-			if (pta[cnt - 1] > pta[cnt])
-			{
-				printf("|%10s | %8d |  i%d | %f | %f | %11.1f | %16s |\n", name, maximum, (int) size * 8, best / 1000000.0, average / 1000000.0, comps, "unstable");
-
-				return;
-			}
-		}
-	}
-
-	if (!strcmp(name, "quadsort"))
-	{
-		if (!strcmp(desc, "random order") || !strcmp(desc, "random 1-4"))
-		{
-			printf("%s\n", "|      Name |    Items | Type |     Best |  Average | Comparisons |     Distribution |");
-			printf("%s\n", "| --------- | -------- | ---- | -------- | -------- | ----------- | ---------------- |");
-		}
-		else
-		{
-			printf("%s\n", "|           |          |      |          |          |             |                  |");
-		}
-	}
-
-	if (counter)
-	{
-		printf("|%10s | %8d |  i%d | %f | %f | %11.1f | %16s |\n", name, maximum, (int) size * 8, best / 1000000.0, average / 1000000.0, comps, desc);
-	}
-	else
-	{
-		printf("|%10s | %8d |  i%d | %f | %f | %11s | %16s |\n", name, maximum, (int) size * 8, best / 1000000.0, average / 1000000.0, "     ", desc);
-	}
-
-	if (minimum != maximum)
-	{
-		return;
-	}
-
-	for (cnt = 1 ; cnt < maximum ; cnt++)
-	{
-		if (size == sizeof(int))
-		{
-			if (pta[cnt - 1] > pta[cnt])
-			{
-				printf("%17s: not properly sorted at index %d. (%d vs %d\n", name, cnt, pta[cnt - 1], pta[cnt]);
-				break;
-			}
-		}
-		else if (size == sizeof(long long))
-		{
-			if (ptla[cnt - 1] > ptla[cnt])
-			{
-				printf("%17s: not properly sorted at index %d. (%016lld vs %016lld\n", name, cnt, ptla[cnt - 1], ptla[cnt]);
-				break;
-			}
-		}
-	}
-
-	for (cnt = 1 ; cnt < maximum ; cnt++)
-	{
-		if (size == sizeof(int))
-		{
-			if (pta[cnt] != ptv[cnt])
-			{
-				printf("         validate: array[%d] != valid[%d]. (%d vs %d\n", cnt, cnt, pta[cnt], ptv[cnt]);
-				break;
-			}
-		}
-		else if (size == sizeof(long long))
-		{
-			if (ptla[cnt] != ptlv[cnt])
-			{
-				printf("         validate: array[%d] != valid[%d]. (%lld vs %lld\n", cnt, cnt, ptla[cnt], ptlv[cnt]);
-				break;
-			}
-		}
-	}
-}
-
-int validate()
-{
-	int seed = time(NULL);
-	int cnt, val, max = 2000000;
-
-	int *a_array, *r_array, *v_array;
-
-	srand(seed);
-
-	a_array = (int *) malloc(max * sizeof(int));
-	r_array = (int *) malloc(max * sizeof(int));
-	v_array = (int *) malloc(max * sizeof(int));
-
-	for (cnt = 0 ; cnt < max ; cnt++)
-	{
-		r_array[cnt] = rand();
-	}
-
-
-	for (cnt = 1 ; cnt < 100 ; cnt++)
-	{
-		memcpy(a_array, r_array, max * sizeof(int));
-		memcpy(v_array, r_array, max * sizeof(int));
-
-		quadsort(a_array, cnt, sizeof(int), cmp_int);
-		qsort(v_array, cnt, sizeof(int), cmp_int);
-
-		for (val = 0 ; val < cnt ; val++)
-		{
-			if (val && v_array[val - 1] > v_array[val])
-			{
-				printf("\e[1;31mvalidate rand: seed %d: size: %d Not properly sorted at index %d.\n", seed, cnt, val);
-				return 0;
-			}
-
-			if (a_array[val] != v_array[val])
-			{
-				printf("\e[1;31mvalidate rand: seed %d: size: %d Not verified at index %d.\n", seed, cnt, val);
-				return 0;
-			}
-		}
-	}
-
-	// ascending saw
-
-	for (cnt = 0 ; cnt < 1000 ; cnt++)
-	{
-		r_array[cnt] = rand();
-	}
-
-        quadsort(r_array + max / 4 * 0, max / 4, sizeof(int), cmp_int);
-        quadsort(r_array + max / 4 * 1, max / 4, sizeof(int), cmp_int);
-        quadsort(r_array + max / 4 * 2, max / 4, sizeof(int), cmp_int);
-        quadsort(r_array + max / 4 * 3, max / 4, sizeof(int), cmp_int);
-
-        for (cnt = 1 ; cnt < 1000 ; cnt += 7)
-	{
-		memcpy(a_array, r_array, max * sizeof(int));
-	        memcpy(v_array, r_array, max * sizeof(int));
-
-	        quadsort(a_array, cnt, sizeof(int), cmp_int);
-	        qsort(v_array, cnt, sizeof(int), cmp_int);
-
-		for (val = 0 ; val < cnt ; val++)
-		{
-			if (val && v_array[val - 1] > v_array[val])
-			{
-				printf("\e[1;31mvalidate ascending saw: seed %d: size: %d Not properly sorted at index %d.\n", seed, cnt, val);
-				return 0;
-			}
-
-			if (a_array[val] != v_array[val])
-			{
-				printf("\e[1;31mvalidate ascending saw: seed %d: size: %d Not verified at index %d.\n", seed, cnt, val);
-				return 0;
-			}
-		}
-	}
-
-        // descending saw
-
-        for (cnt = 0 ; cnt < 1000 ; cnt++)
-        {
-                r_array[cnt] = (max - cnt - 1) % 100000;
-        }
-
-        for (cnt = 1 ; cnt < 1000 ; cnt += 7)
-	{
-		memcpy(a_array, r_array, max * sizeof(int));
-	        memcpy(v_array, r_array, max * sizeof(int));
-
-	        quadsort(a_array, cnt, sizeof(int), cmp_int);
-	        qsort(v_array, cnt, sizeof(int), cmp_int);
-
-		for (val = 0 ; val < cnt ; val++)
-		{
-			if (val && v_array[val - 1] > v_array[val])
-			{
-				printf("\e[1;31mvalidate descending saw: seed %d: size: %d Not properly sorted at index %d.\n", seed, cnt, val);
-				return 0;
-			}
-
-			if (a_array[val] != v_array[val])
-			{
-				printf("\e[1;31mvalidate descending saw: seed %d: size: %d Not verified at index %d.\n", seed, cnt, val);
-				return 0;
-			}
-		}
-	}
-
-	// random tail
-
-	for (cnt = 0 ; cnt < max * 3 / 4 ; cnt++)
-	{
-		r_array[cnt] = cnt;
-	}
-
-	for (cnt = max * 3 / 4 ; cnt < max ; cnt++)
-	{
-		r_array[cnt] = rand();
-	}
-
-        for (cnt = 1 ; cnt < 1000 ; cnt += 7)
-	{
-		memcpy(a_array, r_array, max * sizeof(int));
-	        memcpy(v_array, r_array, max * sizeof(int));
-
-	        quadsort(a_array, cnt, sizeof(int), cmp_int);
-	        qsort(v_array, cnt, sizeof(int), cmp_int);
-
-		for (val = 0 ; val < cnt ; val++)
-		{
-			if (val && v_array[val - 1] > v_array[val])
-			{
-				printf("\e[1;31mvalidate rand tail: seed %d: size: %d Not properly sorted at index %d.\n", seed, cnt, val);
-				return 0;
-			}
-
-			if (a_array[val] != v_array[val])
-			{
-				printf("\e[1;31mvalidate rand tail: seed %d: size: %d Not verified at index %d.\n", seed, cnt, val);
-				return 0;
-			}
-		}
-	}
-
-	free(a_array);
-	free(r_array);
-	free(v_array);
-
-	return 1;
-}
-
-
-int main(int argc, char **argv)
-{
-	int max = 100000;
-	int samples = 10;
-	int repetitions = 1;
-	long long *la_array, *lr_array, *lv_array;
-	int *a_array, *r_array, *v_array;
-	int cnt, rnd;
-
-	if (argc >= 1 && argv[1] && *argv[1])
-	{
-		max = atoi(argv[1]);
-	}
-
-	if (argc >= 2 && argv[2] && *argv[2])
-	{
-		samples = atoi(argv[2]);
-	}
-
-	if (argc >= 3 && argv[3] && *argv[3])
-	{
-		repetitions = atoi(argv[3]);
-	}
-
-	if (argc >4 && argv[4] && *argv[4])
-	{
-		rnd = atoi(argv[4]);
-	}
-	else
-	{
-		rnd = time(NULL);
-	}
-
-	validate();
-
-	printf("benchmark: array size: %d, sample size: %d, repetitions: %d, seed: %d\n\n", max, samples, repetitions, rnd);
-
-
-	// 64 bit
-
-	la_array = (long long *) malloc(max * sizeof(long long));
-	lr_array = (long long *) malloc(max * sizeof(long long));
-	lv_array = (long long *) malloc(max * sizeof(long long));
-
-	if (la_array == NULL || lr_array == NULL || lv_array == NULL)
-	{
-		printf("main(%d,%d,%d): malloc: %s\n", max, samples, repetitions, strerror(errno));
-
-		return 0;
-	}
-
-	// random 
-
-	srand(rnd);
-
-	for (cnt = 0 ; cnt < max ; cnt++)
-	{
-		lr_array[cnt] = rand();
-		lr_array[cnt] += (unsigned long long) rand() << 32ULL;
-	}
-
-	memcpy(lv_array, lr_array, max * sizeof(long long));
-	qsort(lv_array, max, sizeof(long long), cmp_long);
-
-	test_sort(la_array, lr_array, lv_array, max, max, samples, repetitions, quadsort,        "quadsort",       "random order", sizeof(long long), cmp_long);
-	test_sort(la_array, lr_array, lv_array, max, max, samples, repetitions, qsort,           "qsort",          "random order", sizeof(long long), cmp_long);
-
-	printf("\n");
-
-
-	// 32 bit
-
-	a_array = (int *) malloc(max * sizeof(int));
-	r_array = (int *) malloc(max * sizeof(int));
-	v_array = (int *) malloc(max * sizeof(int));
-
-	// random
-
-	srand(rnd);
-
-	for (cnt = 0 ; cnt < max ; cnt++)
-	{
-		r_array[cnt] = rand();
-	}
-
-	memcpy(v_array, r_array, max * sizeof(int));
-	quadsort(v_array, max, sizeof(int), cmp_int);
-
-	test_sort(a_array, r_array, v_array, max, max, samples, repetitions, quadsort,        "quadsort",       "random order", sizeof(int), cmp_int);
-	test_sort(a_array, r_array, v_array, max, max, samples, repetitions, qsort,           "qsort",          "random order", sizeof(int), cmp_int);
-
-	// ascending
-
-	for (cnt = 0 ; cnt < max ; cnt++)
-	{
-		r_array[cnt] = cnt;
-	}
-
-        memcpy(v_array, r_array, max * sizeof(int));
-        memcpy(r_array, v_array, max * sizeof(int));
-
-	quadsort(v_array, max, sizeof(int), cmp_int);
-
-	test_sort(a_array, r_array, v_array, max, max, samples, repetitions, quadsort,        "quadsort",       "ascending", sizeof(int), cmp_int);
-	test_sort(a_array, r_array, v_array, max, max, samples, repetitions, qsort,           "qsort",          "ascending", sizeof(int), cmp_int);
-
-	// ascending saw
-
-	for (cnt = 0 ; cnt < max ; cnt++)
-	{
-		r_array[cnt] = rand();
-	}
-
-        memcpy(v_array, r_array, max * sizeof(int));
-        quadsort(v_array + max / 4 * 0, max / 4, sizeof(int), cmp_int);
-        quadsort(v_array + max / 4 * 1, max / 4, sizeof(int), cmp_int);
-        quadsort(v_array + max / 4 * 2, max / 4, sizeof(int), cmp_int);
-        quadsort(v_array + max / 4 * 3, max / 4, sizeof(int), cmp_int);
-        memcpy(r_array, v_array, max * sizeof(int));
-
-	quadsort(v_array, max, sizeof(int), cmp_int);
-
-	test_sort(a_array, r_array, v_array, max, max, samples, repetitions, quadsort,        "quadsort",       "ascending saw", sizeof(int), cmp_int);
-	test_sort(a_array, r_array, v_array, max, max, samples, repetitions, qsort,           "qsort",          "ascending saw", sizeof(int), cmp_int);
-
-	// uniform
-
-	for (cnt = 0 ; cnt < max ; cnt++)
-	{
-		r_array[cnt] = rand() % 100;
-	}
-
-	memcpy(v_array, r_array, max * sizeof(int));
-	quadsort(v_array, max, sizeof(int), cmp_int);
-
-	test_sort(a_array, r_array, v_array, max, max, samples, repetitions, quadsort,        "quadsort",        "generic order", sizeof(int), cmp_int);
-	test_sort(a_array, r_array, v_array, max, max, samples, repetitions, qsort,           "qsort",           "generic order", sizeof(int), cmp_int);
-
-        // descending
-
-        for (cnt = 0 ; cnt < max ; cnt++)
-        {
-                r_array[cnt] = (max - cnt);
-        }
-
-        memcpy(v_array, r_array, max * sizeof(int));
-        quadsort(v_array, max, sizeof(int), cmp_int);
-
-	test_sort(a_array, r_array, v_array, max, max, samples, repetitions, quadsort,        "quadsort",        "descending order", sizeof(int), cmp_int);
-	test_sort(a_array, r_array, v_array, max, max, samples, repetitions, qsort,           "qsort",           "descending order", sizeof(int), cmp_int);
-
-        // descending saw
-
-        for (cnt = 0 ; cnt < max ; cnt++)
-        {
-                r_array[cnt] = (max - cnt - 1) % 10000;
-        }
-
-        memcpy(v_array, r_array, max * sizeof(int));
-        quadsort(v_array, max, sizeof(int), cmp_int);
-	test_sort(a_array, r_array, v_array, max, max, samples, repetitions, quadsort,        "quadsort",        "descending saw", sizeof(int), cmp_int);
-	test_sort(a_array, r_array, v_array, max, max, samples, repetitions, qsort,           "qsort",           "descending saw", sizeof(int), cmp_int);
-
-	// random tail
-
-	srand(rnd);
-
-	for (cnt = 0 ; cnt < max ; cnt++)
-	{
-		r_array[cnt] = rand();
-	}
-
-	memcpy(v_array, r_array, max * sizeof(int));
-	quadsort(v_array, max * 3 / 4, sizeof(int), cmp_int);
-	memcpy(r_array, v_array, max * sizeof(int));
-	quadsort(v_array, max, sizeof(int), cmp_int);
-
-	test_sort(a_array, r_array, v_array, max, max, samples, repetitions, quadsort,        "quadsort",        "random tail", sizeof(int), cmp_int);
-	test_sort(a_array, r_array, v_array, max, max, samples, repetitions, qsort,           "qsort",           "random tail", sizeof(int), cmp_int);
-
-	// random half
-
-	srand(rnd);
-
-	for (cnt = 0 ; cnt < max ; cnt++)
-	{
-		r_array[cnt] = rand();
-	}
-
-        memcpy(v_array, r_array, max * sizeof(int));
-	quadsort(v_array, max / 2, sizeof(int), cmp_int);
-
-	memcpy(r_array, v_array, max * sizeof(int));
-	quadsort(v_array, max, sizeof(int), cmp_int);
-
-	test_sort(a_array, r_array, v_array, max, max, samples, repetitions, quadsort,        "quadsort",        "random half", sizeof(int), cmp_int);
-	test_sort(a_array, r_array, v_array, max, max, samples, repetitions, qsort,           "qsort",           "random half", sizeof(int), cmp_int);
-
-	// wave
-
-	for (cnt = 0 ; cnt < max ; cnt++)
-	{
-		if (cnt % 2)
-		{
-			r_array[cnt] = 16777216 + cnt;
-		}
-		else
-		{
-			r_array[cnt] = 33554432 + cnt;
-		}
-	}
-
-	memcpy(v_array, r_array, max * sizeof(int));
-	quadsort(v_array, max, sizeof(int), cmp_int);
-
-	test_sort(a_array, r_array, v_array, max, max, samples, repetitions, quadsort,        "quadsort",       "wave order", sizeof(int), cmp_int);
-	test_sort(a_array, r_array, v_array, max, max, samples, repetitions, qsort,           "qsort",          "wave order", sizeof(int), cmp_int);
-
-	test_sort(a_array, r_array, v_array, max, max, samples, repetitions, quadsort,        "quadsort",       "stable", sizeof(int), cmp_stable);
-	test_sort(a_array, r_array, v_array, max, max, samples, repetitions, qsort,           "qsort",          "stable", sizeof(int), cmp_stable);
-
-	if (repetitions > 0)
-	{
-		return 0;
-	}
-
-	// random range
-
-	if (max >= 4)
-	{
-		srand(rnd);
-
-		for (cnt = 0 ; cnt < 4 ; cnt++)
-		{
-			r_array[cnt] = rand();
-		}
-
-		memcpy(v_array, r_array, 8 * sizeof(int));
-		quadsort(v_array, 8, sizeof(int), cmp_int);
-
-		test_sort(a_array, r_array, v_array, 1, 4, 100, 8192, quadsort,        "quadsort",          "random 1-4", sizeof(int), cmp_int);
-		test_sort(a_array, r_array, v_array, 1, 4, 100, 8192, qsort,           "qsort",             "random 1-4", sizeof(int), cmp_int);
-	}
-
-	if (max >= 8)
-	{
-		srand(rnd);
-
-		for (cnt = 0 ; cnt < 8 ; cnt++)
-		{
-			r_array[cnt] = rand();
-		}
-
-		memcpy(v_array, r_array, 8 * sizeof(int));
-		quadsort(v_array, 8, sizeof(int), cmp_int);
-
-		test_sort(a_array, r_array, v_array, 5, 8, 100, 8192, quadsort,        "quadsort",          "random 5-8", sizeof(int), cmp_int);
-		test_sort(a_array, r_array, v_array, 5, 8, 100, 8192, qsort,           "qsort",             "random 5-8", sizeof(int), cmp_int);
-	}
-
-	if (max >= 16)
-	{
-		srand(rnd);
-
-		for (cnt = 0 ; cnt < 16 ; cnt++)
-		{
-			r_array[cnt] = rand();
-		}
-
-		memcpy(v_array, r_array, 16 * sizeof(int));
-		quadsort(v_array, 16, sizeof(int), cmp_int);
-
-		test_sort(a_array, r_array, v_array, 9, 15, 100, 8192, quadsort,        "quadsort",          "random 9-15", sizeof(int), cmp_int);
-		test_sort(a_array, r_array, v_array, 9, 15, 100, 8192, qsort,           "qsort",             "random 9-15", sizeof(int), cmp_int);
-	}
-
-	if (max >= 64)
-	{
-		srand(rnd);
-
-		for (cnt = 0 ; cnt < 64 ; cnt++)
-		{
-			r_array[cnt] = rand();
-		}
-
-		memcpy(v_array, r_array, 64 * sizeof(int));
-		quadsort(v_array, 64, sizeof(int), cmp_int);
-
-		test_sort(a_array, r_array, v_array, 16, 63, 100, 4096, quadsort,        "quadsort",        "random 16-63", sizeof(int), cmp_int);
-		test_sort(a_array, r_array, v_array, 16, 63, 100, 4096, qsort,           "qsort",           "random 16-63", sizeof(int), cmp_int);
-	}
-
-	if (max >= 128)
-	{
-		srand(rnd);
-
-		for (cnt = 0 ; cnt < 128 ; cnt++)
-		{
-			r_array[cnt] = rand();
-		}
-
-		memcpy(v_array, r_array, 128 * sizeof(int));
-		quadsort(v_array, 128, sizeof(int), cmp_int);
-
-		test_sort(a_array, r_array, v_array, 64, 127, 100, 2048, quadsort,        "quadsort",        "random 64-127", sizeof(int), cmp_int);
-		test_sort(a_array, r_array, v_array, 64, 127, 100, 2048, qsort,           "qsort",           "random 64-127", sizeof(int), cmp_int);
-	}
-
-	if (max >= 256)
-	{
-		srand(rnd);
-
-		for (cnt = 0 ; cnt < 256 ; cnt++)
-		{
-			r_array[cnt] = rand();
-		}
-
-		memcpy(v_array, r_array, 256 * sizeof(int));
-		quadsort(v_array, 256, sizeof(int), cmp_int);
-
-		test_sort(a_array, r_array, v_array, 127, 255, 100, 1024, quadsort,        "quadsort",        "random 128-255", sizeof(int), cmp_int);
-		test_sort(a_array, r_array, v_array, 127, 255, 100, 1024, qsort,           "qsort",           "random 128-255", sizeof(int), cmp_int);
-	}
-
-	if (max >= 512)
-	{
-		srand(rnd);
-
-		for (cnt = 0 ; cnt < 512 ; cnt++)
-		{
-			r_array[cnt] = rand();
-		}
-
-		memcpy(v_array, r_array, 512 * sizeof(int));
-		quadsort(v_array, 512, sizeof(int), cmp_int);
-
-		test_sort(a_array, r_array, v_array, 256, 511, 100, 512, quadsort,        "quadsort",        "random 256-511", sizeof(int), cmp_int);
-		test_sort(a_array, r_array, v_array, 256, 511, 100, 512, qsort,           "qsort",           "random 256-511", sizeof(int), cmp_int);
-	}
-
-	if (max >= 1024)
-	{
-		srand(rnd);
-
-		for (cnt = 0 ; cnt < 1024 ; cnt++)
-		{
-			r_array[cnt] = rand();
-		}
-
-		memcpy(v_array, r_array, 1024 * sizeof(int));
-		quadsort(v_array, 1024, sizeof(int), cmp_int);
-
-		test_sort(a_array, r_array, v_array, 512, 1023, 100, 256, quadsort,        "quadsort",        "random 512-1023", sizeof(int), cmp_int);
-		test_sort(a_array, r_array, v_array, 512, 1023, 100, 256, qsort,           "qsort",           "random 512-1023", sizeof(int), cmp_int);
-	}
-
-	if (max >= 2048)
-	{
-		srand(rnd);
-
-		for (cnt = 0 ; cnt < 2048 ; cnt++)
-		{
-			r_array[cnt] = rand();
-		}
-
-		memcpy(v_array, r_array, 2048 * sizeof(int));
-		quadsort(v_array, 2048, sizeof(int), cmp_int);
-
-		test_sort(a_array, r_array, v_array, 1024, 2047, 100, 128, quadsort,        "quadsort",          "random 1024-2047", sizeof(int), cmp_int);
-		test_sort(a_array, r_array, v_array, 1024, 2047, 100, 128, qsort,           "qsort",             "random 1024-2047", sizeof(int), cmp_int);
-	}
-
-	if (max >= 4096)
-	{
-		srand(rnd);
-
-		for (cnt = 0 ; cnt < 4096 ; cnt++)
-		{
-			r_array[cnt] = rand();
-		}
-
-		memcpy(v_array, r_array, 4096 * sizeof(int));
-		quadsort(v_array, 4096, sizeof(int), cmp_int);
-
-		test_sort(a_array, r_array, v_array, 2048, 4095, 100, 64, quadsort,        "quadsort",          "random 2048-4095", sizeof(int), cmp_int);
-		test_sort(a_array, r_array, v_array, 2048, 4095, 100, 64, qsort,           "qsort",             "random 2048-4095", sizeof(int), cmp_int);
-	}
-
-	free(la_array);
-	free(lr_array);
-	free(lv_array);
-
-	free(a_array);
-	free(r_array);
-	free(v_array);
-
-	return 0;
-}
